Move config stuff to a separate package
